code: extract SwitchChainer to utils
